is/requirements.txt
streamlit
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
matplotlib
plotly==6.0.0
joblib==1.4.2
seaborn==0.12.2
tensorflow
Pillow
